edit daily expense kids

+++ database/migrations/2014_10_12_200000_add_two_factor_columns_to_users_table.php
@@ -5,24 +5,26 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 use Laravel\Fortify\Fortify;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->text('two_factor_secret')
+            $table
+                ->text('two_factor_secret')
                 ->after('password')
                 ->nullable();
 
-            $table->text('two_factor_recovery_codes')
+            $table
+                ->text('two_factor_recovery_codes')
                 ->after('two_factor_secret')
                 ->nullable();
 
             if (Fortify::confirmsTwoFactorAuthentication()) {
-                $table->timestamp('two_factor_confirmed_at')
+                $table
+                    ->timestamp('two_factor_confirmed_at')
                     ->after('two_factor_recovery_codes')
                     ->nullable();
             }
@@ -35,12 +37,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn(array_merge([
-                'two_factor_secret',
-                'two_factor_recovery_codes',
-            ], Fortify::confirmsTwoFactorAuthentication() ? [
-                'two_factor_confirmed_at',
-            ] : []));
+            $table->dropColumn(array_merge(['two_factor_secret', 'two_factor_recovery_codes'], Fortify::confirmsTwoFactorAuthentication() ? ['two_factor_confirmed_at'] : []));
         });
     }
 };
